requirements.txt
python-dotenv==0.14.0
tweepy==3.8.0